==> code/03_Amazon_Modeling.ipynb <==
# Amazon Notebook

## Variables

import pandas as pd
import requests
import time
import numpy as np
import matplotlib.pyplot as plt

from nltk.corpus import stopwords

from sklearn.svm import SVC
from sklearn.model_selection import train_test_split, GridSearchCV, RandomizedSearchCV
from sklearn.pipeline import Pipeline
from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import StandardScaler
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import GradientBoostingClassifier, BaggingClassifier,\
                                RandomForestClassifier, ExtraTreesClassifier, AdaBoostClassifier, VotingClassifier

# Reading in the data

df = pd.read_csv('./data/df_export.csv')

df.isna().sum()

# Dropping null values

df = df.dropna()

# Assigning variables

X = df['body']
y = df['subreddit']

# Splitting the data for training and testing

X_train, X_test, y_train, y_test = train_test_split(X, y)

### Data Logging

model_scores = []

## Models

### Simple Logistic Regression

log_pipeline = Pipeline([
    ('tf', TfidfVectorizer()),
    ('logreg', LogisticRegression(solver = 'lbfgs'))
])
params = [{
    'tf__max_features' : [1000, 1250],
    'tf__stop_words'   : [None, 'english'],
    'tf__ngram_range'  : [(1, 1), (1, 2)],
    'logreg__C'         : [.5, 1, 2]
}]
gs_log = GridSearchCV(log_pipeline, param_grid = params, 
                  cv = 5, n_jobs = -1, verbose = 1)

gs_log.fit(X_train, y_train)
log_score = {'estimator' : 'Logistic Regression/Tfidf',
            'train score' : gs_log.score(X_train, y_train),
            'test score' : gs_log.score(X_test, y_test),
            'best params'    : gs_log.best_params_
}
model_scores.append(log_score)
print(log_score)

log_pipeline = Pipeline([
    ('vec', CountVectorizer()),
    ('logreg', LogisticRegression(solver = 'lbfgs'))
])
params = [{
    'vec__max_features' : [1000, 1250],
    'vec__stop_words'   : [None, 'english'],
    'vec__ngram_range'  : [(1, 1), (1, 2)],
    'logreg__C'         : [.5, 1, 2]
}]
gs_log = GridSearchCV(log_pipeline, param_grid = params, 
                  cv = 5, n_jobs = -1, verbose = 1)

gs_log.fit(X_train, y_train)
log_score = {'estimator' : 'Logistic Regression',
            'train score' : gs_log.score(X_train, y_train),
            'test score' : gs_log.score(X_test, y_test),
            'best params'    : gs_log.best_params_
}
model_scores.append(log_score)
print(log_score)

It looks like the test scores for Logistic Regression with both CountVectorizer and TFIDFVectorizer are almost identical. I'm going to continue using CountVectorizer because I am more familiar with its parameters. Also of note, both models liked the highest max_features parameter, so I will increase that hyperparameter to see if it yields better results

### Multinomial Naive Bayes

nb_pipeline = Pipeline([
    ('vec', CountVectorizer()),
    ('nb', MultinomialNB())
])
params = [{
    'vec__max_features' : [1250, 1500],
    'vec__stop_words'   : [None, 'english'],
    'vec__ngram_range'  : [(1, 1), (1, 2)],
    'nb__alpha'         : [.5, 1, 2]
}]
gs_nb = GridSearchCV(nb_pipeline, param_grid = params, 
                  cv = 5, n_jobs = -1, verbose = 2)

gs_nb.fit(X_train, y_train)
nb_score = {'estimator' : 'Multinomial Naive Bayes',
            'train score' : gs_nb.score(X_train, y_train),
            'test score' : gs_nb.score(X_test, y_test),
            'params'    : gs_nb.best_params_
}
model_scores.append(nb_score)
print(nb_score)

Again, the model seemed to do best with the maximum hyperparameter for max_features. I am going to keep increasing it until the performance boost bottoms out. Also of note, all three models performed best with the base ngram range of (1,1). I will continue using that ngram range from here on out.

### Random Forest

rf_pipeline = Pipeline([
    ('vec', CountVectorizer()),
    ('rf', RandomForestClassifier()),
])

params = [{
    'vec__max_features': [1500, 1750],
    'vec__stop_words'  : [None],
    'vec__ngram_range' : [(1, 1)],
    'rf__n_estimators' : [100, 250, 500],
    'rf__max_features': ['auto'],
    'rf__max_depth': [None, 3, 5],
    'rf__min_samples_split' : [2, 3, 5]
}]
gs_rf = GridSearchCV(rf_pipeline, param_grid = params, 
                  cv = 3, n_jobs = -1, verbose = 5)

gs_rf.fit(X_train, y_train)
rf_score = {'estimator' : 'Random Forest',
            'train score' : gs_rf.score(X_train, y_train),
            'test score' : gs_rf.score(X_test, y_test),
            'best params'    : gs_rf.best_params_
}
model_scores.append(rf_score)
print(rf_score)

### Extra Trees

et_pipeline = Pipeline([
    ('vec', CountVectorizer()),
    ('et', ExtraTreesClassifier()),
])

params = [{
    'vec__max_features': [1750, 2000],
    'vec__stop_words'  : [None],
    'vec__ngram_range' : [(1, 1)],
    'et__n_estimators' : [100, 250, 500],
    'et__max_features': ['auto'],
    'et__max_depth': [None, 3, 5], 
    'et__min_samples_split' : [2, 3, 5]
}]
gs_et = GridSearchCV(et_pipeline, param_grid = params, 
                  cv = 3, n_jobs = -1, verbose = 3)

gs_et.fit(X_train, y_train)
et_score = {'estimator' : 'Extra Trees',
            'train score' : gs_et.score(X_train, y_train),
            'test score' : gs_et.score(X_test, y_test),
            'best params' : gs_et.best_params_
}
model_scores.append(et_score)
print(et_score)

### K Nearest Neighbor

knn_pipeline = Pipeline([
    ('vec', CountVectorizer()),
    ('ss', StandardScaler(with_mean = False)),
    ('knn', KNeighborsClassifier())
])

params = [{
    'vec__max_features': [1750, 2000],
    'vec__stop_words'  : [None],
    'vec__ngram_range' : [(1, 1)],
    'knn__n_neighbors' : [3, 4, 5],
    'knn__p'           : [1, 2]
}]
gs_knn = GridSearchCV(knn_pipeline, param_grid = params, 
                  cv = 3, n_jobs = -1, verbose = 5)

gs_knn.fit(X_train, y_train)

knn_score = {'estimator' : 'K Nearest Neighbor',
            'train score' : gs_knn.score(X_train, y_train),
            'test score' : gs_knn.score(X_test, y_test),
            'best params' : gs_knn.best_params_
}
model_scores.append(knn_score)
print(knn_score)

### Bagging Classifier

bc_pipeline = Pipeline([
    ('vec', CountVectorizer()),
    ('bc', BaggingClassifier()),
])

params = [{
    'vec__max_features': [1750, 2000],
    'vec__stop_words'  : [None],
    'vec__ngram_range' : [(1, 1)],
    'bc__n_estimators' : [10, 50, 100]
}]
gs_bc = GridSearchCV(bc_pipeline, param_grid = params, 
                  cv = 3, n_jobs = -1, verbose = 5)

gs_bc.fit(X_train, y_train)

bc_score = {'estimator' : 'Bagging Classifier',
            'train score' : gs_bc.score(X_train, y_train),
            'test score' : gs_bc.score(X_test, y_test),
            'best params' : gs_bc.best_params_
}
model_scores.append(bc_score)
print(bc_score)

### Gradient Booster

gb_pipeline = Pipeline([
    ('vec', CountVectorizer()),
    ('gb', GradientBoostingClassifier())
])

params = [{
    'vec__max_features': [1750, 2000],
    'vec__stop_words'  : [None],
    'vec__ngram_range' : [(1, 1)],
    'gb__n_estimators' : [100, 500, 750],
    'gb__min_samples_split' : [2, 3, 5]
}]
gs_gb = GridSearchCV(gb_pipeline, param_grid = params, 
                  cv = 3, n_jobs = -1, verbose = 2)

gs_gb.fit(X_train, y_train)

gb_score = {'estimator' : 'Gradient Booster',
            'train score' : gs_gb.score(X_train, y_train),
            'test score' : gs_gb.score(X_test, y_test),
            'best params' : gs_gb.best_params_
}
model_scores.append(gb_score)
print(gb_score)

### Support Vector Classifier

svc_pipeline = Pipeline([
    ('vec', CountVectorizer()),
    ('ss', StandardScaler(with_mean = False)),
    ('svc', SVC()),
])

params = [{
    'vec__max_features': [2000],
    'vec__stop_words'  : [None],
    'vec__ngram_range' : [(1, 1)],
    'svc__kernel'      : ['rbf', 'poly'],
    'svc__degree'      : [2],
    'svc__C'           : [.5, 1]
}]
gs_svc = GridSearchCV(svc_pipeline, param_grid = params, 
                  cv = 3, n_jobs = -1, verbose = 5)

gs_svc.fit(X_train, y_train)

svc_score = {'estimator' : 'Support Vector Classifier',
            'train score' : gs_svc.score(X_train, y_train),
            'test score' : gs_svc.score(X_test, y_test),
             'best params' : gs_svc.best_params_
}
model_scores.append(svc_score)
print(svc_score)

## Comprehensive Gridsearch

The random forest, while overfit, performed the best on the test dataset. I will instantiate a random forest model with a more comprehensive gridsearch through its hyperparameters

rf_large_pipeline = Pipeline([
    ('vec', CountVectorizer()),
    ('rf', RandomForestClassifier()),
])

params = [{
    'vec__max_features': [2250, 2500],
    'vec__stop_words'  : [None],
    'vec__ngram_range' : [(1, 1)],
    'rf__n_estimators' : [250, 300],
    'rf__max_features': ['auto', 'log2'],
    'rf__max_depth': [None],
    'rf__min_samples_split' : [4, 5, 6],
    'rf__min_samples_leaf': [1, 2, 3]
    
}]
gs_rf_large = GridSearchCV(rf_large_pipeline, param_grid = params, 
                  cv = 3, n_jobs = 4, verbose = 5)

gs_rf_large.fit(X_train, y_train)

rf_large_score = {'estimator' : 'Large Random Forest',
            'train score' : gs_rf_large.score(X_train, y_train),
            'test score' : gs_rf_large.score(X_test, y_test),
            'best params'    : gs_rf_large.best_params_
}
model_scores.append(rf_large_score)
print(rf_large_score)

## Exporting Results

df_scores = pd.DataFrame(model_scores)
df_scores = df_scores.drop(columns = ['best params', 'params'])

importances = gs_rf_large.best_estimator_['rf'].feature_importances_

important_words = gs_rf_large.best_estimator_.named_steps.vec.get_feature_names()

coefs = pd.DataFrame(importances, important_words)

df_scores.to_csv('./results/scores_dataframe.csv')
coefs.to_csv('./results/word_coefs.csv')
